implemented users and auth

+++ server/src/services/customerService.ts
@@ -2,15 +2,31 @@ import Customer, {
   CustomerDocument,
   CustomerInput,
 } from '../models/customer.model';
+import { UserDocument } from '../models/user.model';
+import logger from '../utils/logger';
 
-export const getAllCustomers = async () => {
-  const customers = await Customer.find<CustomerDocument>({});
-  return customers;
+export const getCustomers = async (user: UserDocument | undefined) => {
+  // const customers = await Customer.find<CustomerDocument>({ user });
+  // return customers;
+  // const user = await User.findById(id).populate('customers');
+  const populated = await user?.populate('customers');
+
+  return populated?.customers;
 };
 
-export const getCustomerById = async (id: string) => {
-  const customer = await Customer.findById<CustomerDocument>(id);
-  return customer;
+export const getSingleCustomer = async (
+  user: UserDocument | undefined,
+  customerId: string
+) => {
+  // const customer = await Customer.findById<CustomerDocument>({id});
+  // return customer;
+
+  const populated = await user?.populate({
+    path: 'customers',
+    match: { _id: customerId },
+  });
+
+  return populated?.customers[0];
 };
 
 // TODO: work out best method of returning the objects
@@ -23,32 +39,56 @@ export const createCustomer = async (data: CustomerInput) => {
   // await customer.save();
   // return customer;
   const customer = await Customer.create<CustomerInput>(data);
-  return customer as CustomerDocument;
+  return customer;
 };
 
-export const deleteCustomerById = async (id: string) => {
+export const deleteCustomerById = async (user: UserDocument, id: string) => {
+  user.customers;
+  const foundCustomer = user.customers.find((c) => c?.toString() === id);
+  if (!foundCustomer) {
+    return null;
+  }
+
   const deleted = await Customer.findOneAndDelete<CustomerDocument>({
     _id: id,
   });
+
   return deleted;
 };
 
-export const editCustomerById = async (id: string, data: object) => {
-  const edited = await Customer.findOneAndReplace({ _id: id }, data, {
+export const editCustomerById = async (
+  user: UserDocument,
+  id: string,
+  data: CustomerInput
+) => {
+  // const customer = await Customer.findOne({ _id: id });
+
+  // logger.info(`editCustomerById: ${JSON.stringify(customer)}`);
+
+  // if (!customer || customer.user.toString() !== user.id) {
+  //   return null;
+  // }
+
+  // customer.overwrite({
+  //   ...data,
+  //   user: user._id
+  // });
+  // await customer.save();
+  logger.info(`editCustomerById: user: ${JSON.stringify(user)}`);
+  // TODO: c is actually of type Types.ObjectId as user.customers is NOT populated - see if mongoose can create a type for this
+  const foundCustomer = user.customers.find((c) => c?.toString() === id);
+  if (!foundCustomer) {
+    return null;
+  }
+
+  const customer = await Customer.findOneAndUpdate({ _id: id }, data, {
     returnDocument: 'after',
   });
-  return edited;
-};
 
-// createCustomer({
-//   name: 'bob',
-//   email: 'bob@hotmail.com',
-//   address: {
-//     line1: '123 park st',
-//     city: 'London',
-//     postcode: 'LN10',
-//   }
-// })
-//   .then(() => console.log(''))
-//   .catch((er) => console.error(er));
+  logger.info(`editCustomerById: after overwrite: ${JSON.stringify(customer)}`);
+  // const edited = await Customer.findOneAndReplace({ _id: id }, data, {
+  //   returnDocument: 'after',
+  // });
+  return customer;
+};
